Rename io.mincongh to io.mincong
