+++ docker/pom.xml
@@ -4,8 +4,8 @@
   xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
   <parent>
+    <groupId>io.mincong</groupId>
     <artifactId>learning-elasticsearch-parent</artifactId>
-    <groupId>io.mincongh</groupId>
     <version>1.0-SNAPSHOT</version>
   </parent>
