+++ test-framework/pom.xml
@@ -5,7 +5,7 @@
   xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
   <modelVersion>4.0.0</modelVersion>
   <parent>
-    <groupId>io.mincongh</groupId>
+    <groupId>io.mincong</groupId>
     <artifactId>learning-elasticsearch-parent</artifactId>
     <version>1.0-SNAPSHOT</version>
   </parent>
