+++ test-framework/src/test/java/io/mincong/elasticsearch/GetTest.java
@@ -1,4 +1,4 @@
-package io.mincongh.elasticsearch;
+package io.mincong.elasticsearch;
 
 import java.util.HashMap;
 import java.util.Map;
