+++ test-framework/src/test/java/io/mincong/elasticsearch/SearchScrollTest.java
@@ -1,4 +1,4 @@
-package io.mincongh.elasticsearch;
+package io.mincong.elasticsearch;
 
 import java.util.ArrayList;
 import java.util.Map;
