+++ test-framework/src/test/java/io/mincong/elasticsearch/SearchTest.java
@@ -1,4 +1,4 @@
-package io.mincongh.elasticsearch;
+package io.mincong.elasticsearch;
 
 import java.util.Map;
 import org.elasticsearch.action.bulk.BulkItemResponse;
